sow: per-file sow, and honor `curable` in per-file ops
The per-file commit only wired up gather. `wildflower sow <path>` ignored
its argument (wildflower.js called `sow()` with no arg, and `sow()` took
no parameter), so it sowed the whole valley — defeating the per-file
workflow in the deploy direction.

Now an explicit path sows only that file/subtree. The per-file resolution
(owning meadow, `if` condition, target restriction, filter/exclude) was
identical to gather's, so it's extracted into a shared `copyPath()` in
common.js parameterized by direction (live→mirror vs mirror→live); gather
and sow both delegate to it. No-arg gather/sow keep the wholesale loops.

`copyPath` routes through `curableCopy` for curable meadows, matching the
non-curable path. `curableCopy` previously ignored the `filter` it was
handed (a latent bug: wholesale curable meadows never honored their
include/exclude rules either) — it now applies the filter via a shared
`matchesFilter()` helper, so curable meadows respect `!*-tokens.json` and
friends in both wholesale and per-file modes.
